requirements.txt
cs50==9.4.0
Flask==3.1.0
Flask_Caching==2.3.0
Flask_JWT_Extended==4.7.1
python-dotenv==1.0.1
waitress==3.0.2
